dont use properties on toplevel
