+++ samples/ControlCatalog/MainWindow.xaml
@@ -5,10 +5,7 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:vm="clr-namespace:ControlCatalog.ViewModels"
         xmlns:v="clr-namespace:ControlCatalog.Views"
-        x:Class="ControlCatalog.MainWindow" 
-        LocalNotificationManager="{Binding NotificationManager, Mode=OneWayToSource}" 
-        NotificationArea.Position="TopRight"
-        NotificationArea.MaxItems="3">
+        x:Class="ControlCatalog.MainWindow">
     <Window.DataTemplates>
         <DataTemplate DataType="vm:NotificationViewModel">
             <v:NotificationView />
